separate build commands for individual packages
