+++ .travis.yml
@@ -1,6 +1,6 @@
 language: node_js
 script:
-  - yarn bolt
+  - yarn build
   - yarn test
   - yarn flow
   - yarn lint
