+++ packages/codesandboxer-fs/index.js
@@ -1,13 +1,11 @@
 // @flow
-const assembleFiles = require('./assembleFiles');
-const assembleFilesAndPost = require('./assembleFilesAndPost');
+const assembleFiles = require('./src/assembleFiles');
+const assembleFilesAndPost = require('./src/assembleFilesAndPost');
 
 /* eslint-disable-next-line */
 module.exports = { assembleFilesAndPost, assembleFiles };
 
 // other stuff we need to handle:
 // loading nonJS files
-//   .jsx
-//   .png +
 //   .tsx
 // Check we never reach above our initial scope
