+++ packages/codesandboxer-fs/src/assembleFiles.js
